requirements.txt
boto3==1.20.14
botocore==1.23.14
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
idna==3.3
jmespath==0.10.0
joblib==1.1.0
nltk==3.6.5
numpy==1.21.4
Pillow==8.4.0
pyenchant==3.2.2
pytesseract==0.3.8
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
regex==2021.11.10
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
torch==1.10.0
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
